split large test files up
